Add config `preserve-script-closures` as method/event closure work-around.
Converting JavaScript functions to the serializable `Handler` format
will break closures as the function passes through the kernel and is
translated in and out of a `Handler`. This is necessary to be able to
replicate methods and event listeners that components conjure up at
runtime.

However, it may break applications that inadvertently depend on
references to internal variables visible to certain functions. These
applications can work for a single client (although they would break
in any case when the application is replicated to a second client).

Setting `preserve-script-closures` will allow these applications to
continue to work for the limited case of a single client.